Mutual fund Classification
A mutual fund is a professionally managed investment fund that pools money
from many investors to purchase securities. These investors may be retail or
institutional in nature. Mutual fund ratings is one of the most influential and
prominent decision making factors that is used by investors in making a
decision with regards to selecting a mutual fund.
Great Stone Rating is a star based ranking system.
These ratings are based on the performance of a mutual
fund with adjustments for risks and costs as compared to
other funds in the same category. The rating ranges from
0 to 5.
Goal: The goal of this hackathon is to predict GreatStone’s
rating of a mutual fund. In order to help investors decide
on which mutual fund to pick for an investment, the task is
to build a model that can predict the rating of a mutual
fund. The various attributes that define a mutual fund can
be used for building the model.